fix: Change category rendering to server-side
