--- src/main/java/com/woowahan/techcamp/recipehub/category/controller/CategoryRestController.java
@@ -1,28 +1,0 @@
-package com.woowahan.techcamp.recipehub.category.controller;
-
-import com.woowahan.techcamp.recipehub.category.domain.Category;
-import com.woowahan.techcamp.recipehub.category.service.CategoryService;
-import com.woowahan.techcamp.recipehub.common.support.RestResponse;
-import com.woowahan.techcamp.recipehub.recipe.service.RecipeService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/api/categories")
-public class CategoryRestController {
-
-    @Autowired
-    private CategoryService categoryService;
-
-    @Autowired
-    private RecipeService recipeService;
-
-    @GetMapping
-    public RestResponse<List<Category>> getAllCategories() {
-        return RestResponse.success(categoryService.findAll());
-    }
-}
