+++ src/main/resources/templates/index.hbs
@@ -5,7 +5,11 @@
 
 {{#partial "contents"}}
     <div id="category-nav">
-        <div id="category-box" class="container level menu-list"></div>
+        <div id="category-box" class="container level menu-list">
+            {{#each categories}}
+                <p class="category has-text-centered column child-vertical-align" data-category-id={{id}}>{{title}}</p>
+            {{/each}}
+        </div>
     </div>
 <section class="section">
     <div class="container">
