--- src/test/java/com/woowahan/techcamp/recipehub/category/controller/CategoryRestAcceptanceTest.java
@@ -1,48 +1,0 @@
-package com.woowahan.techcamp.recipehub.category.controller;
-
-import com.woowahan.techcamp.recipehub.category.domain.Category;
-import com.woowahan.techcamp.recipehub.category.repository.CategoryRepository;
-import com.woowahan.techcamp.recipehub.common.support.RestResponse;
-import com.woowahan.techcamp.recipehub.recipe.repository.RecipeRepository;
-import com.woowahan.techcamp.recipehub.support.AcceptanceTest;
-import org.junit.Before;
-import org.junit.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.ParameterizedTypeReference;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-
-import java.util.List;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-public class CategoryRestAcceptanceTest extends AcceptanceTest {
-
-    @Autowired
-    private CategoryRepository categoryRepository;
-
-    @Autowired
-    private RecipeRepository recipeRepository;
-
-    private Category categoryA;
-    private Category categoryB;
-
-    @Override
-    @Before
-    public void setUp() throws Exception {
-        super.setUp();
-        categoryA = categoryRepository.save(Category.builder().title("A").build());
-        categoryB = categoryRepository.save(Category.builder().title("B").build());
-    }
-
-    @Test
-    public void readAllCategories() {
-        ResponseEntity<RestResponse<List<Category>>> response = requestJson("/api/categories", HttpMethod.GET, new ParameterizedTypeReference<RestResponse<List<Category>>>() {
-        });
-
-        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
-        assertThat(response.getBody().getData()).contains(categoryA, categoryB);
-    }
-
-}
