+++ src/test/java/com/woowahan/techcamp/recipehub/home/HomeAcceptanceTest.java
@@ -6,6 +6,7 @@ import com.woowahan.techcamp.recipehub.recipe.domain.Recipe;
 import com.woowahan.techcamp.recipehub.recipe.repository.RecipeRepository;
 import com.woowahan.techcamp.recipehub.support.AcceptanceTest;
 import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -21,9 +22,21 @@ public class HomeAcceptanceTest extends AcceptanceTest {
     @Autowired
     private CategoryRepository categoryRepository;
 
+    private Category categoryA;
+    private Category categoryB;
+
+    @Override
+    @Before
+    public void setUp() throws Exception {
+        super.setUp();
+        categoryA = categoryRepository.save(Category.builder().title("A").build());
+        categoryB = categoryRepository.save(Category.builder().title("B").build());
+    }
+
     @Test
     public void mainPage() throws Exception {
-        assertThat(requestGet("/").getStatusCode()).isEqualTo(HttpStatus.OK);
+        ResponseEntity<String> response = requestGet("/");
+        assertGetCategory(response.getBody());
     }
 
     @Test
@@ -36,6 +49,12 @@ public class HomeAcceptanceTest extends AcceptanceTest {
         ResponseEntity<String> response = requestGet("/");
         assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
         assertThat(response.getBody()).contains("Hello");
+        assertGetCategory(response.getBody());
+    }
+
+    private void assertGetCategory(String body) {
+        assertThat(body).contains("A");
+        assertThat(body).contains("B");
     }
 
     @Override
